fix bugs users

+++ users/views.py
@@ -1,6 +1,8 @@
 from django.views.generic import TemplateView
 from django.shortcuts import render
 from campus.models import Campus
+from reservas.models import Reserva
+from datetime import datetime
 from .models import *
 from .forms import *
 
@@ -14,6 +16,8 @@ def view_coordenadoresCurso(request):
             data = {}
             form = UserCreationForm(request.POST or None)
             form.fields['email'].required = True
+            form.fields['first_name'].required = True
+            form.fields['last_name'].required = True
             campus = Campus.objects.all()
             if len(campus) > 0:
                 data['campus'] = campus[0]
@@ -45,13 +49,12 @@ def ver_coordenadorCurso(request, id_coordenadorcurso):
             coordenadorCurso = User.objects.get(id=id_coordenadorcurso)
             if coordenadorCurso:
                 data['coordenadorcurso'] = coordenadorCurso
-                return render(request,
-                              'users/coordenadorEnsino/view_coordenadorcurso.html', data)
+                return render(request, 'users/coordenadorEnsino/view_coordenadorcurso.html', data)
             else:
                 data = {'mensagem': "Não foi possível localizar o Coordenador de Curso!"}
                 return render(request, 'users/coordenadorEnsino/error.html', data)
         else:
-            return render(request, '../../templates/permission_error.html')
+            return render(request, 'permission_error.html')
     except:
         data = {'mensagem': "Não foi possível visualizar o coordenador de curso!"}
         return render(request, 'users/coordenadorEnsino/error.html', data)
@@ -65,26 +68,14 @@ def delete_coordenadorCurso(request, id_coordenadorcurso):
         if request.user.tipo_usuario == 'CoordenadorEnsino':
             data = {'mensagem': "Coordenador de Curso " + str(id_coordenadorcurso) + " removido com sucesso!"}
             coordenadorCurso = User.objects.get(id=id_coordenadorcurso)
-            n = coordenadorCurso.delete()
-            print("Removidos: ", n)
-            return render(request, 'users/coordenadorEnsino/cadastro_sucesso.html', data)
-        else:
-            return render(request, '../../templates/permission_error.html')
-    except:
-        data = {'mensagem': "Não foi possível excluir o Coordenador de Curso!"}
-        return render(request, 'users/coordenadorEnsino/error.html', data)
+            reservas = Reserva.objects.filter(user=coordenadorCurso, dataFim__gte=datetime.today().strftime("%Y-%m-%d"))
 
-
-def delete_coordenadorCurso(request, id_coordenadorcurso):
-    try:
-        if not request.user.is_authenticated:
-            return render(request, 'permission_error.html')
-
-        if request.user.tipo_usuario == 'CoordenadorEnsino':
-            data = {'mensagem': "Coordenador de Curso " + str(id_coordenadorcurso) + " removido com sucesso!"}
-            coordenadorCurso = User.objects.get(id=id_coordenadorcurso)
-            coordenadorCurso.delete()
-            return render(request, 'users/coordenadorEnsino/cadastro_sucesso.html', data)
+            if len(reservas) == 0:
+                coordenadorCurso.delete()
+                return render(request, 'users/coordenadorEnsino/cadastro_sucesso.html', data)
+            else:
+                data = {'mensagem': "Não foi possível excluir o Coordenador de Curso! Ele possui reservas em aberto."}
+                return render(request, 'users/coordenadorEnsino/error.html', data)
         else:
             return render(request, 'permission_error.html')
     except:
@@ -101,6 +92,8 @@ def view_professores(request):
             data = {}
             form = UserCreationForm(request.POST or None)
             form.fields['email'].required = True
+            form.fields['first_name'].required = True
+            form.fields['last_name'].required = True
             campus = Campus.objects.all()
             if len(campus) > 0:
                 data['campus'] = campus[0]
@@ -116,7 +109,7 @@ def view_professores(request):
             data['form'] = form
             return render(request, 'users/coordenadorCurso/professores.html', data)
         else:
-            return render(request, '.permission_error.html')
+            return render(request, 'permission_error.html')
     except:
         data = {'mensagem': "Ocorreu um erro interno!" }
         return render(request, 'users/coordenadorCurso/error.html', data)
@@ -149,10 +142,15 @@ def delete_professor(request, id_professor):
             return render(request, 'permission_error.html')
 
         if request.user.tipo_usuario == 'CoordenadorCurso':
-            data = {'mensagem': "Professor " + str(id_professor) + " removido com sucesso!"}
             professor = User.objects.get(id=id_professor)
-            professor.delete()
-            return render(request, 'users/coordenadorCurso/cadastro_sucesso.html', data)
+            reservas = Reserva.objects.filter(user=professor, dataFim__gte=datetime.today().strftime("%Y-%m-%d"))
+            if len(reservas) == 0:
+                data = {'mensagem': "Professor " + professor.nome + " removido com sucesso!"}
+                professor.delete()
+                return render(request, 'users/coordenadorCurso/cadastro_sucesso.html', data)
+            else:
+                data = {'mensagem': "Não foi possível excluir o Professor(a)! Há reservas em aberto."}
+                return render(request, 'users/coordenadorCurso/error.html', data)
         else:
             return render(request, 'permission_error.html')
     except:
